+++ src/DependencyInjection/HeimrichHannotTwigTemplatesExtension.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\TwigTemplatesBundle\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
+use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
+
+class HeimrichHannotTwigTemplatesExtension extends Extension implements PrependExtensionInterface
+{
+    public function load(array $configs, ContainerBuilder $container)
+    {
+        // TODO: Implement load() method.
+    }
+
+    public function prepend(ContainerBuilder $container)
+    {
+        $container->prependExtensionConfig('huh_twig_support', ['enable_template_loader' => true]);
+        $container->prependExtensionConfig('twig', ['paths' => [
+            '%kernel.project_dir%/vendor/heimrichhannot/contao-twig-templates-bundle/src/Resources/views' => 'ContaoTwigTemplates',
+        ]]);
+    }
+}