+++ src/HeimrichHannotTwigTemplatesBundle.php
@@ -9,13 +9,21 @@
 namespace HeimrichHannot\TwigTemplatesBundle;
 
 use HeimrichHannot\TwigTemplatesBundle\DependencyInjection\Compiler\FrontendFrameworkPass;
+use HeimrichHannot\TwigTemplatesBundle\DependencyInjection\HeimrichHannotTwigTemplatesExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class ContaoTwigTemplatesBundle extends Bundle
+class HeimrichHannotTwigTemplatesBundle extends Bundle
 {
     public function build(ContainerBuilder $container)
     {
         $container->addCompilerPass(new FrontendFrameworkPass());
     }
+
+    public function getContainerExtension()
+    {
+        return new HeimrichHannotTwigTemplatesExtension();
+    }
+
+
 }