+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.0-beta8] - 2020-09-22
+- **BREAKING:** renamed bundle class to HeimrichHannotTwigTemplatesBundle (old twig namespaces are preserved, but should be updated till next major version)
+- raised twig support bundle dependency to 0.2
+
 ## [2.0.0-beta7] - 2020-09-17
 - twig template base logic moved into contao-twig-support-bundle
 - removed deprecated code: AbstractTemplate, AbstractFramework, ...